Fix: Class "UserService" has 24 methods, which is greater than 20 authorized. Split it into smaller classes.
